\name{RcmdrPlugin.BCA-package}
\alias{RcmdrPlugin.BCA-package}
\alias{RcmdrPlugin.BCA}
\docType{package}
\title{
RcmdrPlugin.BCA
}
\description{
An R Commander plug-in to accompany the book 
\cite{Customer and Business Analytics: Applied Data Mining for Business Decision Making Using R} by Daniel S. Putler 
and Robert E. Krider.
}
\details{
Package:  RcmdrPlugin.BCA\cr
Version:  0.9-3\cr
Date:  2012/01/28\cr
Depends: R (>= 2.1.0), Rcmdr (>= 1.6-0), tcltk, BCA\cr
Suggests: bind, aplpack, car (>= 1.1-1), clv, effects (>= 1.0-7), foreign, grid, 
  lattice, lmtest, MASS, mgcv, multcomp, nlme, nnet, relimp, rpart, rpart.plot,
  rgl, RODBC\cr
LazyLoad: no\cr
License: GPL (>= 2)\cr
URL: http://www.r-project.org,
http://www.customeranalyticsbook.com/\cr
}

\author{
Dan Putler <putler@yahoo.com>
Maintainer: Dan Putler <putler@yahoo.com>
}

\keyword{ package }
