\name{RcmdrPlugin.BCA-package}
\alias{RcmdrPlugin.BCA-package}
\alias{RcmdrPlugin.BCA}
\docType{package}
\title{
RcmdrPlugin.BCA
}
\description{
An R Commander plug-in to accompany the book 
\cite{Customer and Business Analytics: Applied Data Mining for Business Decision Making Using R} by Daniel S. Putler 
and Robert E. Krider.
}
\details{
Package:  RcmdrPlugin.BCA\cr
Version:  0.9-8\cr
Date:  2013/09/01\cr
Depends: R (>= 3.0.0), BCA, car\cr
Imports: Rcmdr (>= 1.9-4), tcltk\cr
Suggests: bind, aplpack, clv (>- 2.0-15, effects (>= 1.0-7), foreign, grid, 
  lattice, lmtest, MASS, mgcv, multcomp, nlme, nnet, relimp, rpart, rpart.plot,
  rgl, RODBC\cr
LazyLoad: no\cr
License: GPL (>= 2)\cr
URL: http://www.customeranalyticsbook.com/\cr
}

\author{
Dan Putler <putler@yahoo.com>
Maintainer: Dan Putler <putler@yahoo.com>
}

\keyword{ package }
